pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unibase-agent-framework"
version = "0.1.0"
description = "A transparent SDK for building AI agents with integrated memory management"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Unibase Team", email = "team@unibase.io"}
]
keywords = ["ai", "agents", "memory", "langchain", "claude", "openai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "unibase-aip-sdk",
]

[project.optional-dependencies]
claude = ["anthropic>=0.18.0"]
openai = ["openai>=1.0.0"]
langchain = ["langchain>=0.1.0", "langchain-anthropic>=0.1.0"]
mem0 = ["mem0ai>=0.1.0"]
web3 = ["web3>=6.0.0", "eth-account>=0.9.0"]
crypto = ["cryptography>=41.0.0"]
all = [
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "mem0ai>=0.1.0",
    "web3>=6.0.0",
    "eth-account>=0.9.0",
    "cryptography>=41.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/unibaseio/unibase-agent-framework"
Documentation = "https://docs.unibase.io/framework"
Repository = "https://github.com/unibaseio/unibase-agent-framework"

[tool.setuptools.packages.find]
where = ["."]
include = ["unibase_agent_sdk*"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
